Move capture service into its own file
